Added hardcoded profile creation
Users of the Eatseebot will now be able to create a user profile by
entering their age and gender. This is via the profile conversation
and profile handler in main.py, with the methods called defined in
the Profile.py Profile class.

The home page of the Profile section will also show whether the user has
a profile etc, or display his existing profile