Fixed Bugs in Profile, Added General Functions
1. Bug Fix:

Bug detected where eatsee_db API will return preferences
as a list of integers. However, when choices are called from the api,
they are returned as string keys. e.g. 1 vs '1'

This led to discrepancy and errors occured since the eatsee_bot code
assumed them to be the same representation and used them interchangably.
E.g in Profile where choices are selected

Solution:
In class Service, the response from the api for preferences will be
converted to a list a string.

2. Added General Functions

There was alot of duplication in the Profile Component Class. Cleaned
that up by introducing general/Higher order functions. Achieved by
using eval/string formatting. Cons of this is that we will need to make
sure that all the naming convention is consistent, else it will fail.